# common-platform-contract-mgmt-sdk/src/main/resources/api-spec/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "common-platform-contract-mgmt",
    "description": "Contract Management Core Application",
    "contact": {
      "name": "Firefly Software Solutions Inc",
      "email": "dev@getfirefly.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://core.getfirefly.io/contracts",
      "description": "Development Environment"
    },
    {
      "url": "/",
      "description": "Local Development Environment"
    }
  ],
...
  "paths": {
...
    "/api/v1/contracts/{contractId}/parties/filter": {
      "post": {
        "tags": [
          "Contract Parties"
        ],
        "summary": "Filter contract parties",
        "description": "Returns a paginated list of contract parties based on filter criteria",
        "operationId": "filterContractParties",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestContractPartyDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid filter criteria provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contract parties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contracts/filter": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Filter contracts",
        "description": "Returns a paginated list of contracts based on filter criteria",
        "operationId": "filterContracts",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestContractDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid filter criteria provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contracts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contract-parties/filter": {
      "post": {
        "tags": [
          "Global Contract Parties"
        ],
        "summary": "Filter contract parties globally",
        "description": "Returns a paginated list of contract parties based on filter criteria across all contracts",
        "operationId": "filterContractPartiesGlobally",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestContractPartyDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid filter criteria provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contract parties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contract-parties": {
      "get": {
        "tags": [
          "Global Contract Parties"
        ],
        "summary": "Get contract parties by party ID",
        "description": "Returns a paginated list of all contract parties for a specific party. This endpoint is useful for finding all contracts a party is involved in.",
        "operationId": "getContractPartiesByPartyId",
        "parameters": [
          {
            "name": "partyId",
            "in": "query",
            "description": "ID of the party to retrieve contract parties for",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "Filter for active contracts only",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved contract parties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponseContractPartyDTO"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "400": {
            "description": "Invalid party ID provided"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ContractPartyDTO": {
        "required": [
          "contractId",
          "partyId",
          "roleInContractId"
        ],
        "type": "object",
        "properties": {
          "contractPartyId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "contractId": {
            "type": "string",
            "format": "uuid"
          },
          "partyId": {
            "type": "string",
            "format": "uuid"
          },
          "roleInContractId": {
            "type": "string",
            "format": "uuid"
          },
          "dateJoined": {
            "type": "string",
            "format": "date-time"
          },
          "dateLeft": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "ContractDTO": {
        "required": [
          "contractStatus"
        ],
        "type": "object",
        "properties": {
          "contractId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "contractNumber": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "contractStatus": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_APPROVAL",
              "ACTIVE",
              "SUSPENDED",
              "TERMINATED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "productCatalogId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "FilterOptions": {
        "type": "object",
        "properties": {
          "caseInsensitiveStrings": {
            "type": "boolean"
          },
          "includeInheritedFields": {
            "type": "boolean"
          }
        },
        "description": "Filter options for customizing filter behavior"
      },
...
      "PaginationRequest": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "description": "The zero-based page number to retrieve.",
            "format": "int32",
            "example": 0,
            "default": 0
          },
          "pageSize": {
            "type": "integer",
            "description": "The number of items per page.",
            "format": "int32",
            "example": 10,
            "default": 10
          },
          "sortBy": {
            "type": "string",
            "description": "The field to sort the results by.",
            "example": "name"
          },
          "sortDirection": {
            "type": "string",
            "description": "The direction of sorting, either ASC or DESC.",
            "example": "DESC",
            "default": "DESC"
          }
        },
        "description": "Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction."
      },
      "RangeFilter": {
        "type": "object",
        "properties": {
          "ranges": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/RangeObject"
            },
            "description": "Map of field names to their range criteria"
          }
        },
        "description": "Represents range filters for numeric fields"
      },
      "RangeObject": {
        "type": "object",
        "properties": {
          "from": {
            "type": "object",
            "description": "Start value of the range"
          },
          "to": {
            "type": "object",
            "description": "End value of the range"
          }
        },
        "description": "Represents a range with from and to values"
      },
      "PaginationResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "The list of items for the current page.",
            "items": {
              "type": "object",
              "description": "The list of items for the current page."
            }
          },
          "totalElements": {
            "type": "integer",
            "description": "The total number of elements across all pages.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages based on the data size and page size.",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "description": "The current page number, typically zero-based.",
            "format": "int32"
          }
        },
        "description": "Represents a paginated response containing a list of items and pagination metadata."
      },
...
      "FilterRequestContractPartyDTO": {
        "required": [
          "pagination"
        ],
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ContractPartyDTO"
          },
          "rangeFilters": {
            "$ref": "#/components/schemas/RangeFilter"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequest"
          },
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          }
        },
        "description": "Generic filter request that includes both filter criteria and pagination"
      },
...
      "FilterRequestContractDTO": {
        "required": [
          "pagination"
        ],
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ContractDTO"
          },
          "rangeFilters": {
            "$ref": "#/components/schemas/RangeFilter"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequest"
          },
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          }
        },
        "description": "Generic filter request that includes both filter criteria and pagination"
      },
...
      "PaginationResponseContractPartyDTO": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "The list of items for the current page.",
            "items": {
              "$ref": "#/components/schemas/ContractPartyDTO"
            }
          },
          "totalElements": {
            "type": "integer",
            "description": "The total number of elements across all pages.",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "description": "The total number of pages based on the data size and page size.",
            "format": "int32"
          },
          "currentPage": {
            "type": "integer",
            "description": "The current page number, typically zero-based.",
            "format": "int32"
          }
        },
        "description": "Represents a paginated response containing a list of items and pagination metadata."
      }
    }
  }
}